stored data into realm

+++ app/src/main/java/Models/GSONModel/GUser.java
@@ -4,6 +4,6 @@ public class GUser {
 
     public int id;
     public String name, email, password;
-    public int phone;
+    public long phone;
     public int type;
 }
